+++ src/stdlib.rs
@@ -1,1 +1,3 @@
 pub mod math;
+pub mod list;
+pub mod string;
